requirements.txt
amqp==5.3.1
attrs==20.3.0
Authlib==1.3.1
billiard==4.2.1
boto3==1.17.22
botocore==1.20.22
celery==5.4.0
certifi==2024.7.4
cffi==1.14.5
chardet==4.0.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.2.0
cryptography==44.0.1
defusedxml==0.7.0
Django==4.2.20
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-encrypted-model-fields==0.6.5
django-timezone-field==7.1
djangorestframework==3.15.2
factory-boy==3.2.0
Faker==8.1.3
gevent==23.9.1
greenlet==3.0.1
gunicorn==23.0.0
idna==3.7
iniconfig==1.1.1
jira==3.4.0
jira-cache==0.8.3
jmespath==0.10.0
jsonfield==3.1.0
kombu==5.5.1
matplotlib==3.3.4
networkx==2.8.7
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
pbr==5.5.1
pluggy==0.13.1
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pycurl==7.43.0.6
pygit2==1.5.0
PyJWT==2.4.0
pyparsing==2.4.7
pytest==7.2.0
pytest-django==4.1.0
python-crontab==2.5.1
python-dateutil==2.9.0.post0
pytz==2021.1
requests==2.32.2
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.3.4
scipy==1.6.1
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.5.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.2.0